refactor: use types from @signalk/server-api
Use the types that the server API provides where possible.

There is one FIXME for a force cast, where the server's
types are incorrect.

Updates containing values may but usually do not contain meta.
I suggest reconsidering if meta should be stored or not - mostly
meta should be static data, that is a bit orthogonal to storing
values history.